Renamed `CentralProtocol` to `CentralManager`
