--- Sources/GATT/TimeInterval.swift
@@ -1,14 +1,0 @@
-//
-//  TimeInterval.swift
-//  GATT
-//
-//  Created by Alsey Coleman Miller on 7/15/18.
-//
-
-import Foundation
-
-public extension TimeInterval {
-    
-    /// GATT default timeout. 
-    static var gattDefaultTimeout: TimeInterval { return 30.0 }
-}
